Utilities

MethodOfLines.unitindex — Function

unitindex(N, j)

Get a unit CartesianIndex in dimension j of length N.

MethodOfLines.params — Function

params(u, s::DiscreteSpace)

Fillter out the time variable and get the spatial variables of u in s.

MethodOfLines.Idx — Function

Idx(II::CartesianIndex, s::DiscreteSpace, u, indexmap)

Here indexmap maps the arguments of u in s to the their ordering. Return a subindex of II that corresponds to only the spatial arguments of u.