Utilities

Missing docstring.

Missing docstring for unitindex. Check Documenter's build log for details.

MethodOfLines.ivs — Function

ivs(u, s::DiscreteSpace)

Filter out the time variable and get the spatial variables of u in s.

MethodOfLines.Idx — Function

Idx(II::CartesianIndex, s::DiscreteSpace, u, indexmap)

Here indexmap maps the arguments of u in s to their ordering. Return a subindex of II that corresponds to only the spatial arguments of u.